Move azure operation operation to a new folder
