+++ .changeset/good-moles-kneel.md
@@ -1,0 +1,5 @@
+---
+"@azure-tools/cadl-ranch-specs": patch
+---
+
+Move azure page operation to a new folder
