+++ packages/cadl-ranch-specs/http/azure/core/basic/mockapi.ts
@@ -74,53 +74,6 @@ Scenarios.Azure_Core_Basic_list = passOnSuccess(
   }),
 );
 
-Scenarios.Azure_Core_Basic_listWithPage = passOnSuccess(
-  mockapi.get("/azure/core/basic/page", (req) => {
-    const responseBody = {
-      value: [validUser],
-    };
-    return { status: 200, body: json(responseBody) };
-  }),
-);
-
-Scenarios.Azure_Core_Basic_listWithParameters = passOnSuccess(
-  mockapi.get("/azure/core/basic/parameters", (req) => {
-    req.expect.containsQueryParam("another", "Second");
-
-    const validBody = { inputName: "Madge" };
-    req.expect.bodyEquals(validBody);
-
-    const responseBody = {
-      value: [validUser],
-    };
-    return { status: 200, body: json(responseBody) };
-  }),
-);
-
-Scenarios.Azure_Core_Basic_TwoModelsAsPageItem = passOnSuccess([
-  mockapi.get("/azure/core/basic/first-item", () => {
-    const responseBody = {
-      value: [{ id: 1 }],
-    };
-    return { status: 200, body: json(responseBody) };
-  }),
-  mockapi.get("/azure/core/basic/second-item", () => {
-    const responseBody = {
-      value: [{ name: "Madge" }],
-    };
-    return { status: 200, body: json(responseBody) };
-  }),
-]);
-
-Scenarios.Azure_Core_Basic_listWithCustomPageModel = passOnSuccess(
-  mockapi.get("/azure/core/basic/custom-page", () => {
-    const responseBody = {
-      items: [validUser],
-    };
-    return { status: 200, body: json(responseBody) };
-  }),
-);
-
 Scenarios.Azure_Core_Basic_delete = passOnSuccess(
   mockapi.delete("/azure/core/basic/users/:id", (req) => {
     if (req.params.id !== "1") {
